Changes
Tried loading GOMenu after a couple of seconds of being caught (in
Spotted.cs).
Added boxes on one part of richareapavement so footsteps work, however
disabled them because guy jumps when stepping on them for the first
time. Temporarily replaced with box collider on richareapavement for
testers to be able to play.
Edited build setting so MainMenu and GOMenu are included properly. Both
load MainScene22, if new MainScene is made then MainMenu and GOMenu
scripts need to be changed to call the new version of the scene.